Added StackActivity, which contains all operations for a stack + a StackView.
This StackView is the visual representation of the stack.

The user can currently push a number onto the stack (max = INTEGER_MAX / min = -INTEGER_MAX) and pop it. These operations are already animated (but still need a bit of tweaking). The boxes will resize according to the amount of data and the size of the input numbers.
The used Icons are NOT final - the toolbar was inserted for defining the right proportions and can be changed